VB-927: Fix issue where closed visits don't display on visits page if there are no open visits
