+++ server/services/visitSessionsService.ts
@@ -16,9 +16,10 @@ import VisitSchedulerApiClient from '../data/visitSchedulerApiClient'
 import WhereaboutsApiClient from '../data/whereaboutsApiClient'
 import { VisitSession, Visit, SupportType, OutcomeDto } from '../data/visitSchedulerApiTypes'
 import { ScheduledEvent } from '../data/whereaboutsApiTypes'
-import { prisonerDateTimePretty, prisonerTimePretty, sortByTimestamp } from '../utils/utils'
+import { prisonerDateTimePretty, prisonerTimePretty } from '../utils/utils'
 import PrisonerContactRegistryApiClient from '../data/prisonerContactRegistryApiClient'
 import buildVisitorListItem from '../utils/visitorUtils'
+import getVisitSlotsFromBookedVisits from '../utils/visitsUtils'
 import { getSupportTypeDescriptions } from '../routes/visitorUtils'
 
 type PrisonerContactRegistryApiClientBuilder = (token: string) => PrisonerContactRegistryApiClient
@@ -263,68 +264,7 @@ export default class VisitSessionsService {
 
     return {
       extendedVisitsInfo,
-      slots: this.getVisitSlotsFromBookedVisits(extendedVisitsInfo),
-    }
-  }
-
-  private getVisitSlotsFromBookedVisits(visits: ExtendedVisitInformation[]): {
-    openSlots: VisitsPageSlot[]
-    closedSlots: VisitsPageSlot[]
-    firstSlotTime: string
-  } {
-    const openSlots: VisitsPageSlot[] = []
-    const closedSlots: VisitsPageSlot[] = []
-
-    visits.forEach((visit: ExtendedVisitInformation) => {
-      if (visit.visitRestriction === 'OPEN') {
-        let matchingOpenSlot = openSlots.findIndex(openSlot => openSlot.visitTime === visit.visitTime)
-
-        if (matchingOpenSlot < 0) {
-          openSlots.push({
-            visitTime: visit.visitTime,
-            visitType: visit.visitRestriction,
-            sortField: visit.startTimestamp,
-            adults: 0,
-            children: 0,
-          })
-
-          matchingOpenSlot = openSlots.length - 1
-        }
-
-        openSlots[matchingOpenSlot].adults += visit.visitors.filter(visitor => visitor.adult).length
-        openSlots[matchingOpenSlot].children += visit.visitors.filter(visitor => !visitor.adult).length
-      } else {
-        let matchingClosedSlot = closedSlots.findIndex(closedSlot => closedSlot.visitTime === visit.visitTime)
-
-        if (matchingClosedSlot < 0) {
-          closedSlots.push({
-            visitTime: visit.visitTime,
-            visitType: visit.visitRestriction,
-            sortField: visit.startTimestamp,
-            adults: 0,
-            children: 0,
-          })
-
-          matchingClosedSlot = closedSlots.length - 1
-        }
-
-        closedSlots[matchingClosedSlot].adults += visit.visitors.filter(visitor => visitor.adult).length
-        closedSlots[matchingClosedSlot].children += visit.visitors.filter(visitor => !visitor.adult).length
-      }
-    })
-
-    let firstSlotTime: string
-
-    if (openSlots.length > 0) {
-      firstSlotTime = openSlots.sort(sortByTimestamp)[0].visitTime
-    } else if (closedSlots.length > 0) {
-      firstSlotTime = closedSlots.sort(sortByTimestamp)[0].visitTime
-    }
-
-    return {
-      openSlots,
-      closedSlots,
-      firstSlotTime,
+      slots: getVisitSlotsFromBookedVisits(extendedVisitsInfo),
     }
   }
 
@@ -389,7 +329,7 @@ export default class VisitSessionsService {
       startTimestamp: visit.startTimestamp,
       visitDate: prisonerDateTimePretty(visit.startTimestamp),
       visitTime,
-      visitRestriction: visit.visitRestriction,
+      visitRestriction: visit.visitRestriction === 'OPEN' ? 'OPEN' : 'CLOSED',
       visitors,
     }
   }
